refactor: extract isValidEmail to @/lib/email to keep client bundle thin
